third_party/WebKit/LayoutTests/editing/input/change-style-with-key-binding-expected.txt

Test key bindings for toggling bold and italic.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS root.querySelectorAll("b").length is 0
PASS root.querySelectorAll("i").length is 0
PASS root.querySelectorAll("b").length is 1
PASS root.querySelectorAll("i").length is 0
PASS root.querySelectorAll("b").length is 1
PASS root.querySelectorAll("i").length is 1
PASS root.querySelectorAll("b").length is 0
PASS root.querySelectorAll("i").length is 1
PASS root.querySelectorAll("b").length is 0
PASS root.querySelectorAll("i").length is 0
PASS successfullyParsed is true

TEST COMPLETE

Foo
